Close to the finish line

+++ AtmosphericEffects/AtmosphericEffects/Utils.cpp
@@ -1,5 +1,5 @@
 #include "Utils.hpp"
 
 int FrameCounter::tickindex = 0;
-int FrameCounter::ticksum = 0;
-int FrameCounter::ticklist[Samples] = { 0 };
+float FrameCounter::ticksum = 0;
+float FrameCounter::ticklist[Samples] = { 0 };
